Refactor modules directory to only contain games
I.e., rename directory to `games` and move election stage out of
directory.
