+++ lit/src/experiment-components/games/README.md
@@ -1,0 +1,3 @@
+# Experiment Games
+
+This directory contains experiment games (i.e., a set os associated stages).
